Application is ready. Code partially cleaned

+++ src/main/java/keray/ui/PickUserUI.java
@@ -4,6 +4,7 @@ import javafx.scene.control.*;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
+import keray.domain.Person;
 import keray.logic.DbUsers;
 import keray.logic.UsersTable;
 
@@ -11,25 +12,17 @@ import java.util.Optional;
 
 //UI with the menu to pick user
 public class PickUserUI {
+    BorderPane mainPickUserUI = new BorderPane();
 
     public BorderPane getPickUserUI() {
 
-        BorderPane mainPickUserUI = new BorderPane();
-
-        //top menu
-        VBox topMenu = new VBox();
-
-        //command
+        //command/instruction
         Label commandLabel = new Label("Pick user:");
 
         //new table to pick users
         UsersTable usersTable = new UsersTable();
-        TableView table = usersTable.createNewTable();
-
-        table.setColumnResizePolicy(TableView.CONSTRAINED_RESIZE_POLICY);
-
-        table.prefWidthProperty().bind(ParentUI.getParentLayout().widthProperty().add(-50));
-        table.prefHeightProperty().bind(ParentUI.getParentLayout().heightProperty().multiply(0.35));
+        TableView<Person> table = usersTable.createNewTable();
+        this.formatUserTable(table);
 
         //Region object to make a space between table and delete button
         Region space = new Region();
@@ -37,44 +30,62 @@ public class PickUserUI {
 
         //Delete user button
         Button deleteButton = new Button("Delete User");
-        deleteButton.setPrefHeight(40);
-        deleteButton.prefWidthProperty().bind(ParentUI.getParentLayout().widthProperty().add(-50));
+        this.formatButton(deleteButton);
 
-        //adding elements to the top menu
+        //Creating topMenu to segregate nodes at the top of UI
+        VBox topMenu = new VBox();
         topMenu.getChildren().addAll(commandLabel, table, space, deleteButton);
+
 
         //Button to add new user
         Button addUserButton = new Button("Add new user");
-        addUserButton.setPrefHeight(40);
-        addUserButton.prefWidthProperty().bind(ParentUI.getParentLayout().widthProperty().add(-50));
+        this.formatButton(addUserButton);
 
         //adding all elements to the layout
-        mainPickUserUI.setTop(topMenu);
-        mainPickUserUI.setBottom(addUserButton);
+        this.mainPickUserUI.setTop(topMenu);
+        this.mainPickUserUI.setBottom(addUserButton);
+
+        ////////////ANIMATING UI///////////////////////////////////
 
         //Deleting user action
         deleteButton.setOnAction((event) -> {
             if (usersTable.toDelete != null) {
                 this.userDeleteConfirmationDialog(usersTable, table);
             } else {
-                this.userNotPickedDialog();
+                this.userNotSelectedDialog();
             }
         });
 
         //Adding new user action
         addUserButton.setOnAction((event) -> {
             AddUserUI addingUserLayout = new AddUserUI();
-            VBox addingUser = addingUserLayout.addUserUI();
+            VBox addingUser = addingUserLayout.getAddUserUI();
             ParentUI.getParentLayout().getChildren().removeAll(topMenu, addUserButton);
             ParentUI.setInsideParentLayout(addingUser);
         });
 
-        return mainPickUserUI;
+        return this.mainPickUserUI;
+    }
+
+    ////////////////////METHODS///////////////////////////////
+
+    //Method will format delete user button
+    private void formatButton(Button button) {
+        button.setPrefHeight(40);
+        button.prefWidthProperty().bind(ParentUI.getParentLayout().widthProperty().add(-50));
+    }
+
+    //Method will adjust user table
+    private void formatUserTable(TableView<Person> table) {
+        table.setColumnResizePolicy(TableView.CONSTRAINED_RESIZE_POLICY);
+
+        table.prefWidthProperty().bind(ParentUI.getParentLayout().widthProperty().add(-50));
+        table.prefHeightProperty().bind(ParentUI.getParentLayout().heightProperty().multiply(0.35));
     }
 
 
-    public void userDeleteConfirmationDialog(UsersTable userstable, TableView table) {
-        Dialog<ButtonType> dialog = new Dialog<ButtonType>();
+    private void userDeleteConfirmationDialog(UsersTable userstable, TableView<Person> table) {
+        Dialog<ButtonType> dialog = new Dialog<>();
         dialog.setTitle("Confirm");
         ButtonType yesButton = new ButtonType("Yes");
         ButtonType noButton = new ButtonType("No");
@@ -85,7 +96,7 @@ public class PickUserUI {
         Optional<ButtonType> result = dialog.showAndWait();
 
         //deleting record after user confirms
-        if (result.get() == yesButton && userstable.toDelete != null) {
+        if (result.isPresent() && result.get() == yesButton && userstable.toDelete != null) {
             DbUsers.deleteUserFromDb(userstable.toDelete);
             userstable.toDelete = null;
             userstable.updateTable(table);
@@ -95,14 +106,13 @@ public class PickUserUI {
     }
 
     //dialog informing about user not being picked
-    public void userNotPickedDialog() {
-        Dialog<ButtonType> dialog = new Dialog<ButtonType>();
+    private void userNotSelectedDialog() {
+        Dialog<ButtonType> dialog = new Dialog<>();
         dialog.setTitle("Select user to delete");
         ButtonType yesButton = new ButtonType("Ok");
         dialog.setContentText("User to delete is not selected");
         dialog.getDialogPane().getButtonTypes().add(yesButton);
-
-        Optional<ButtonType> result = dialog.showAndWait();
+        dialog.showAndWait();
     }
